requirements.txt
streamlit
pandas
numpy
plotly
supabase
qrcode
Pillow
fpdf
requests
pdfplumber
reportlab
matplotlib
seaborn
google-generativeai
xlsxwriter
python-dotenv